Move endpoints out of index.js
